Def päivitys ja asetyypin päivitys esinettä vaihdettaessa

+++ Assets/1Scripts/Items/Equipment/Wearable/Armor.cs
@@ -5,9 +5,9 @@ using UnityEngine;
 [CreateAssetMenu(fileName = "New armor", menuName = "Item/Equipment/Armor")]
 public class Armor : Equipment
 {
-    public float defence;
+    public int defence;
 
-    public Armor(string itemName, float weight, float defence, Sprite inventoryPicture) : base(itemName, inventoryPicture, weight)
+    public Armor(string itemName, float weight, int defence, Sprite inventoryPicture) : base(itemName, inventoryPicture, weight)
     {
         this.defence = defence;
     }
